# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexo App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GoogleLoginRequest {
  google_id: string;
  email: string;
  name?: string;
}

export interface AuthResponse {
  account_id: number;
  email: string;
  onboarding_completed: boolean;
}

export interface GoalRanking {
  goal_id: number;
  rank: number;
}

export interface OnboardingRequest {
  account_id: number;
  name: string;
  age: number;
  gender: string;
  height_cm: number;
  weight_kg: number;
  phone_number: string;
  goals: GoalRanking[];
}

export interface OnboardingResponse {
  success: boolean;
  profile_id?: number;
  message: string;
}

export interface Profile {
  id: number;
  account_id: number;
  name: string;
  age: number;
  gender: string;
  height_cm: number;
  weight_kg: number;
  bmi: number;
  phone_number: string;
}

export interface Recommendation {
  product_id: number;
  product_name: string;
  grams: number;
  grams_per_day: number;
  goal_name: string;
}

export interface DashboardResponse {
  profile: Profile;
  recommendations: Recommendation[];
  total_grams_per_month: number;
  grams_per_day: number;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price_per_gram: number;
  goal_id: number;
  goal_name: string;
  is_trial: boolean;
  trial_price?: number;
}

export interface ProductsResponse {
  products: Product[];
}

export interface AddToCartRequest {
  account_id: number;
  product_id: number;
  grams: number;
}

export interface CartItem {
  id: number;
  account_id: number;
  product_id: number;
  product_name: string;
  grams: number;
  price: number;
}

export interface CartResponse {
  items: CartItem[];
  total_price: number;
}

export interface CheckoutRequest {
  account_id: number;
}

export interface OrderItem {
  id: number;
  product_id: number;
  product_name: string;
  grams: number;
  price: number;
}

export interface Order {
  id: number;
  account_id: number;
  total_price: number;
  status: string;
  tracking_link?: string;
  created_at: string;
  items: OrderItem[];
}

export interface OrderResponse {
  order: Order;
  message: string;
}

export interface OrdersResponse {
  orders: Order[];
}

export interface SupportResponse {
  phone: string;
  email: string;
}

export interface MessageResponse {
  message: string;
}

export type GetDashboardParams = {
  account_id: number;
};

export type GetCartParams = {
  account_id: number;
};

export type GetOrdersParams = {
  account_id: number;
};