# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexo App API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";

export interface Order {
  id: number;
  account_id: number;
  total_price: number;
  status: string;
  tracking_link?: string;
  created_at: string;
  items: OrderItem[];
}