# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexo App API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  id: number;
  product_id: number;
  product_name: string;
  grams: number;
  price: number;
}