# lib/api-zod/src/generated/types/orderResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexo App API
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from "./order";

export interface OrderResponse {
  order: Order;
  message: string;
}