# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexo App API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price_per_gram: number;
  goal_id: number;
  goal_name: string;
  is_trial: boolean;
  trial_price?: number;
}