# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nexo App API
 * OpenAPI spec version: 0.1.0
 */

export interface Recommendation {
  product_id: number;
  product_name: string;
  grams: number;
  grams_per_day: number;
  goal_name: string;
}